+++ src/main/java/com/workintech/s17d2/tax/Taxable.java
@@ -1,7 +1,7 @@
 package com.workintech.s17d2.tax;
 
 public interface Taxable {
-    int getSimpleTaxRate();
-    int getMiddleTaxRate();
-    int getUpperTaxRate();
+    Double getSimpleTaxRate();
+    Double getMiddleTaxRate();
+    Double getUpperTaxRate();
 }
